fixed the issue with failing tests; changed the validator behaviour
